refactor(polyfill): extract ipfs, i18n and RN into files
