+++ src/node.polyfill.ts
@@ -4,10 +4,6 @@
  * backendMiddleware and redux store under node
  */
 
-function logCall(pref: string) {
-  return (msg: string) => console.log(`${pref}: ${msg}`)
-}
-
 import { BackendMiddleware } from 'src/backendMiddleware'
 import { createStore, applyMiddleware } from 'redux'
 import thunk from 'redux-thunk'
@@ -36,31 +32,4 @@ export function initStore() {
   }
 
   return store
-}
-
-export {
-  IpfsStorageAgent as IpfsCustomConnector,
-} from 'node_modules/jolocom-lib/js/ipfs/ipfs'
-
-// default export merge of all default exports
-export default {
-  // src/locales/i18n.ts default export
-  t: (str: string) => str,
-
-  // react-native-splash-screen
-  hide: () => logCall('SplashScreen')('hide'),
-}
-
-import { randomBytes } from 'crypto'
-export const NativeModules = { RNRandomBytes: { randomBytes } }
-
-export class Linking {
-  static async canOpenURL(url: string) {
-    console.log('canOpenURL called with ' + url)
-    return true
-  }
-
-  static async openURL(url: string) {
-    console.log('openURL called with ' + url)
-  }
 }
