+++ src/polyfills/ipfs.ts
@@ -1,0 +1,3 @@
+export {
+  IpfsStorageAgent as IpfsCustomConnector,
+} from 'jolocom-lib/js/ipfs/ipfs'
