+++ tsconfig.json
@@ -46,10 +46,10 @@
       "src/lib/keychain": ["./src/polyfills/keychain"],
       "rn-fetch-blob": ["./src/polyfills/rnFetchBlob"],
       "src/lib/errors/sentry": ["./src/polyfills/sentry"],
-      "src/locales/i18n": ["./src/node.polyfill"],
-      "react-native-splash-screen": ["./src/node.polyfill"],
-      "react-native": ["./src/node.polyfill"],
-      "src/lib/ipfs": ["./src/node.polyfill"],
+      "src/locales/i18n": ["./src/polyfills/i18n"],
+      "react-native-splash-screen": ["./src/polyfills/rnSplashScreen"],
+      "react-native": ["./src/polyfills/reactNative"],
+      "src/lib/ipfs": ["./src/polyfills/ipfs"],
       "src/*": ["./src/*"],
       "typeorm/browser": ["./../node_modules/typeorm"]
     },
